Update test_ome_template
The compiler now produces a SyntaxWarning when identity checks (is and is not) are used with certain types of literals (e.g. strings, numbers). These can often work by accident in CPython, but are not guaranteed by the language spec. The warning advises users to use equality tests (== and !=) instead.

Co-Authored-By: Sajna Shetty <42166139+Sajna-Shetty@users.noreply.github.com>
Co-Authored-By: Felix Stephen <38071645+felixs88@users.noreply.github.com>